requirements.txt
openai
wikipedia-api
python-dotenv
requests
